разбиение на модули, webpack
