+++ .eslintrc.yml
@@ -1,6 +1,10 @@
+parserOptions:
+  sourceType: 'module'
+  ecmaVersion: 2015
+
 env:
   es6: true
   browser: true
-  commonjs: true
+
 
 extends: 'htmlacademy/es6'
