+++ src/view/board-container.js
@@ -1,0 +1,6 @@
+export const createBoardContainerHtml = () => {
+  return (
+    `<section class="board container">
+    </section>`
+  );
+};
