+++ src/view/bord-task-list.js
@@ -1,0 +1,6 @@
+export const createBordTaskListHtml = () => {
+  return (
+    `<div class="board__tasks">
+    </div>`
+  );
+};
