+++ src/view/show-more-btn.js
@@ -1,0 +1,6 @@
+export
+const createShowMoreBtnHtml = () => {
+  return (
+    `<button class="load-more" type="button">load more</button>`
+  );
+};
